refactor: topic result files re-organization
