+++ web/beacon-app/src/features/topics/components/TopicQueryResult/MimeTypeResult/index.ts
@@ -1,0 +1,1 @@
+export { default as BinaryResult } from './BinaryResult';
